Redesign of the eligibility warning page
